Adding a refresh button to invalidate changelogs cache
